+++ MOVEit.MIT/Public/Get-MITFile.ps1
@@ -4,10 +4,10 @@ function Get-MITFile {
         Get MOVEit Transfer File(s)
     .LINK
         Get list of files current user can view
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/files?Page={Page}&PerPage={PerPage}&SortField={SortField}&SortDirection={SortDirection}&NewOnly={NewOnly}-1.0
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/#operation/GETapi/v1/files?Page={Page}&PerPage={PerPage}&SortField={SortField}&SortDirection={SortDirection}&NewOnly={NewOnly}-1.0
     .LINK
         Return file details        
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/files/{Id}-1.0 
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/#operation/GETapi/v1/files/{Id}-1.0 
     #>
     [CmdletBinding(DefaultParameterSetName='List')]
     param (
@@ -20,6 +20,8 @@ function Get-MITFile {
 
         [Parameter(Mandatory=$false,
                     ParameterSetName='List')]
+        [Parameter(Mandatory=$false,
+                    ParameterSetName='ListAll')]
         [switch]$NewOnly,
 
         [Parameter(Mandatory=$false,
@@ -32,17 +34,25 @@ function Get-MITFile {
 
         [Parameter(Mandatory=$false,
                     ParameterSetName='List')]
+        [Parameter(Mandatory=$false,
+                    ParameterSetName='ListAll')]
         [ValidateSet('name', 'size', 'datetime', 'uploadstamp', 'path')]
         [string]$SortField,
 
         [Parameter(Mandatory=$false,
                     ParameterSetName='List')]
+        [Parameter(Mandatory=$false,
+                    ParameterSetName='ListAll')]
         [ValidateSet('ascending', 'descending')]
         [string]$SortDirection,
         
         [Parameter(Mandatory=$false,
                     ParameterSetName='List')]        
-        [switch]$IncludePaging
+        [switch]$IncludePaging,
+
+        [Parameter(Mandatory=$true,
+                    ParameterSetName='ListAll')]        
+        [switch]$All
     )
     try {
         # Set the resource for this request
@@ -66,6 +76,9 @@ function Get-MITFile {
                 }
                 Invoke-MITRequest -Resource $resource -Body $query |
                     Write-MITResponse -Typename 'MITFileSimple' -IncludePaging:$IncludePaging
+            }
+            'ListAll' {                
+                Invoke-MITGetAll -Scriptblock ${function:Get-MITFile} -BoundParameters $PSBoundParameters
             }
         }
     }