+++ MOVEit.MIT/Public/Get-MITGroupMember.ps1
@@ -4,9 +4,9 @@ function Get-MITGroupMember {
         Get MOVEit Transfer Group Member(s)
     .LINK
         Get a list of the members of a group
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/groups/{Id}/members?Page={Page}&PerPage={PerPage}-1.0        
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/#operation/GETapi/v1/groups/{Id}/members?Page={Page}&PerPage={PerPage}-1.0        
     #>
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName='List')]
     param (
         [Parameter(Mandatory,
                     Position=0,
@@ -14,30 +14,40 @@ function Get-MITGroupMember {
         [Alias('Id')]                    
         [string]$GroupId,
 
-        [Parameter()]
+        [Parameter(ParameterSetName='List')]
         [int32]$Page,
 
-        [Parameter()]
+        [Parameter(ParameterSetName='List')]
         [int32]$PerPage,
 
-        [Parameter()]
-        [switch]$IncludePaging
+        [Parameter(ParameterSetName='List')]
+        [switch]$IncludePaging,
+
+        [Parameter(Mandatory, ParameterSetName='ListAll')]
+        [switch]$All
     )  
     
     try {
         # Set the resource for this request
         $resource = "groups/$GroupId/members"
-                    
-        # Build the query parameters.
-        $query = @{}
-        switch ($PSBoundParameters.Keys) {
-            Page { $query['page'] = $Page }
-            PerPage { $query['perPage'] = $PerPage }
-        }
+        
+        switch ($PSCmdlet.ParameterSetName) {
+            'List' {
+                # Build the query parameters.
+                $query = @{}
+                switch ($PSBoundParameters.Keys) {
+                    Page { $query['page'] = $Page }
+                    PerPage { $query['perPage'] = $PerPage }
+                }
 
-        # Send the request and write out the response
-        Invoke-MITRequest -Resource "$resource" -Body $query |
-            Write-MITResponse -Typename 'MITUserSimple' -IncludePaging:$IncludePaging
+                # Send the request and write out the response
+                Invoke-MITRequest -Resource "$resource" -Body $query |
+                    Write-MITResponse -Typename 'MITUserSimple' -IncludePaging:$IncludePaging
+            }
+            'ListAll' {
+                Invoke-MITGetAll -Scriptblock ${function:Get-MITGroupMember} -BoundParameters $PSBoundParameters
+            }
+        }
     }
     catch {
         $PSCmdlet.ThrowTerminatingError($PSItem)