+++ MOVEit.MIT/Public/Get-MITInfo.ps1
@@ -4,7 +4,7 @@ function Get-MITInfo {
         Get MOVEit Transfer server public org info
     .LINK
         Gets public information about organization
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/info?orgId={orgId}-1.0        
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/#operation/GETapi/v1/info?orgId={orgId}-1.0        
     #>
     [CmdletBinding()]
     param (