+++ MOVEit.MIT/Public/Get-MITLog.ps1
@@ -39,10 +39,10 @@ function Get-MITLog {
         Paging info as custom MITPaging object
     .LINK
         Get list of logs current user can view
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/index.html#operation/GETapi/v1/logs?Page={Page}&PerPage={PerPage}&SortField={SortField}&SortDirection={SortDirection}&StartDateTime={StartDateTime}&EndDateTime={EndDateTime}&Action={Action}&UserNameContains={UserNameContains}&UserId={UserId}&FileIdContains={FileIdContains}&FileNameContains={FileNameContains}&SizeComparison={SizeComparison}&Size={Size}&FolderId={FolderId}&FolderPathContains={FolderPathContains}&IpContains={IpContains}&AgentBrandContains={AgentBrandContains}&SuccessFailure={SuccessFailure}&SuppressSigns={SuppressSigns}&SuppressEmailNotes={SuppressEmailNotes}&SuppressLogViews={SuppressLogViews}-1.0
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/index.html#operation/GETapi/v1/logs?Page={Page}&PerPage={PerPage}&SortField={SortField}&SortDirection={SortDirection}&StartDateTime={StartDateTime}&EndDateTime={EndDateTime}&Action={Action}&UserNameContains={UserNameContains}&UserId={UserId}&FileIdContains={FileIdContains}&FileNameContains={FileNameContains}&SizeComparison={SizeComparison}&Size={Size}&FolderId={FolderId}&FolderPathContains={FolderPathContains}&IpContains={IpContains}&AgentBrandContains={AgentBrandContains}&SuccessFailure={SuccessFailure}&SuppressSigns={SuppressSigns}&SuppressEmailNotes={SuppressEmailNotes}&SuppressLogViews={SuppressLogViews}-1.0
     .LINK        
         Get log's info
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/index.html#operation/GETapi/v1/logs/{Id}-1.0
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/index.html#operation/GETapi/v1/logs/{Id}-1.0
     #>
     [CmdletBinding()]
     param (           
@@ -56,77 +56,94 @@ function Get-MITLog {
 
         # startDateTime for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [datetime]$StartDateTime,
         
         # endDateTime for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [datetime]$EndDateTime,
 
         # action for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [ValidateSet('None','FileTransfer','Administration','Upload',
                     'Download','UserMaintenance','ContentScanning')]
         [string]$Action,
 
         # userNameContains for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$UserNameContains,
 
         # userId for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$UserId,
 
         # fileIdContains for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$FileIdContains,
 
         # fileNameContains for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$FileNameContains,
 
         # sizeComparison for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [ValidateSet('None','LargerOrEqual','SmallerOrEqual')]
         [string]$SizeComparison,
 
         # size for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$Size,
 
         # folderId for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$FolderId,
 
         # folderPathContains for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$FolderPathContains,
         
         # ipContains for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$IpContains,
 
         # agentBrandContains for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [string]$AgentBrandContains,
 
         # successFailure for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [ValidateSet('None', 'Success', 'Error')]
         [string]$SuccessFailure,
 
         # suppressSigns for REST call
         # Note this switch sets suppressSigns to False
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [switch]$IncludeSigns,
 
         # suppressEmailNotes for REST call
         # Note this switch sets suppressEmailNotes to False
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [switch]$IncludeEmailNotes,
 
         # suppressLogViews for REST call
         # Note this switch sets suppressLogViews to False
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [switch]$IncludeLogViews,        
         
         # page for REST call
@@ -139,6 +156,7 @@ function Get-MITLog {
 
         # sortField for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [ValidateSet('id', 'logtime', 'action', 'username', 'userrealname', 'targetname', 
                     'filename', 'fileid', 'folderpath', 'xfersize', 'duration', 
                     'rate', 'ipaddress', 'agentbrand', 'resilnode')]
@@ -146,12 +164,16 @@ function Get-MITLog {
         
         # sortDirection for REST call
         [Parameter(Mandatory=$false, ParameterSetName='List')]
+        [Parameter(Mandatory=$false, ParameterSetName='ListAll')]
         [ValidateSet('asc', 'desc')]
         [string]$SortDirection,
         
         # switch to not include PagingInfo in the output
         [Parameter(Mandatory=$false)]
-        [switch]$IncludePaging
+        [switch]$IncludePaging,
+
+        [Parameter(Mandatory, ParameterSetName='ListAll')]
+        [switch]$All        
     )
     
     try {
@@ -198,6 +220,9 @@ function Get-MITLog {
                 # Send the request and write out the response
                 Invoke-MITRequest -Resource $resource -Body $query |
                     Write-MITResponse -TypeName 'MITLog' -IncludePaging:$IncludePaging
+            }
+            'ListAll' {
+                Invoke-MITGetAll -Scriptblock ${function:Get-MITLog} -BoundParameters $PSBoundParameters
             }
         }
     }