+++ MOVEit.MIT/Public/New-MITUser.ps1
@@ -4,7 +4,7 @@ function New-MITUser {
         Create a MOVEit Transfer User
     .LINK
         Create new user
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/POSTapi/v1/users-1.0        
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/#operation/POSTapi/v1/users-1.0        
     #>
     [CmdletBinding()]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingPlainTextForPassword', 'Password')]