+++ MOVEit.MIT/Public/Set-MITUser.ps1
@@ -4,7 +4,7 @@ function Set-MITUser {
         Change a MOVEit Transfer User
     .LINK
         Partial user update
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/PATCHapi/v1/users/{Id}-1.0        
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/#operation/PATCHapi/v1/users/{Id}-1.0        
     #>
     [CmdletBinding()]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingPlainTextForPassword', 'Password')]