+++ MOVEit.MIT/Public/Switch-MITOrg.ps1
@@ -4,7 +4,7 @@ function Switch-MITOrg {
         Switch (ie. ActAsAdmin) to a different Org.
     .LINK
         Become administrator in specific organization
-        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/POSTapi/v1/auth/actasadmin-1.0        
+        https://docs.ipswitch.com/MOVEit/Transfer2023/Api/Rest/#operation/POSTapi/v1/auth/actasadmin-1.0        
     #>
     [CmdletBinding()]
     param (